Add files via upload
CircleFun is a Python program that uses the Turtle module to draw a circle over and around another circle. The output of this program is a visually appealing design that is created by the combination of these two circles.
The program first sets up the Turtle module by importing it and creating a turtle object. Then, it uses the penup() and pendown() methods to move the turtle to the center of the canvas and draw the first circle using the circle() method. The size of this circle can be adjusted by changing the radius parameter of the circle() method.
Next, the turtle is moved to the edge of the first circle and rotated using the setheading() method so that it faces outwards. The turtle then draws another circle using the circle() method, but this time with a larger radius. This creates a circle that is over the first circle.
Finally, the turtle is moved back to the center of the canvas and rotated again to face upwards. The turtle then writes some text on the canvas using the write() method. This text can be customized by changing the text parameter and font parameters of the write() method.
When the program is run, the output is a visually pleasing design consisting of a circle over and around another circle, with text in the center. The program can be modified to create different designs by adjusting the size and position of the circles, as well as the text.